requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
emoji==2.15.0
executing==2.2.1
faiss-cpu==1.13.2
fastjsonschema==2.21.2
filelock==3.20.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
olefile==0.47
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pillow==12.1.0
pipreqs==0.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.7
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
streamlit==1.53.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.27
unstructured-client==0.42.6
urllib3==2.6.3
uuid_utils==0.13.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==2.0.1
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zstandard==0.25.0
